feat: Introduce dictionary feature and enhance navigation
Adds a new `DICTIONARY` game mode and corresponding `Dictionary` component. This expands the application's learning capabilities beyond flashcards and grammar exercises.

Also, this commit refactors the `CategorySelection` component to accept an `initialCategory` prop, allowing for pre-selection of categories. The `Navbar` component is updated to include the new dictionary option and the "Custom Vocab" label is clarified to "Moji Izrazi". The audio playback in `ActiveGame` is optimized by using a singleton `AudioContext` and implementing a smoother sound envelope for better user experience. The `FlashCard` interface is updated to include an optional `display` property for future use. Finally, user storage interactions are updated to include `addMastered` alongside `addMistakes`.